pyproject.toml
[project]
name = "gmail-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain==0.2.11",
  "gradio==4.39.0",
  "chromadb==0.5.5",
  "google-api-python-client==2.124.0",
  "google-auth==2.35.0",
  "google-auth-oauthlib==1.2.0",
  "python-dotenv==1.0.1",
  "openai==1.51.2",
  "onnxruntime==1.16.3",
  "langchain-openai==0.1.22",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
